feat(agreements): move to api layer
