+++ packages/api/graphql/src/rsc.ts
@@ -4,3 +4,4 @@ export * from './query/dao/list';
 export * from './query/dao/detail';
 export * from './query/dao/user-daos';
 export * from './fetch-access-token';
+export * from './query/agreements/list';
